Fix A* heuristic penalty saturation by upgrading scores to u32
The A* pathfinding algorithm (`get_boss_path` and `astar_search`)
used `u16` for its scores. This caused heuristic penalties to quickly
saturate (at 65,535), preventing the algorithm from correctly
distinguishing between highly dangerous paths (e.g. fatal obstacles)
and moderately dangerous paths.

Upgrading the `f_score` to `u32` and scaling the penalties appropriately
resolves the saturation issue and ensures accurate pathfinding.

+++ src/game/a_star_state.rs
@@ -1,6 +1,6 @@
 #[derive(Copy, Clone, Eq, PartialEq)]
 pub struct AStarState {
-    pub f_score: u16,
+    pub f_score: u32,
     pub tie_breaker: u64,
     pub position: crate::snake::Point,
 }
